@@ -8,6 +8,8 @@
 # Generated by versioneer-0.28
 # https://github.com/python-versioneer/python-versioneer
 
+# ruff: noqa: UP031
+
 """Git implementation of _version.py."""
 
 import errno
@@ -16,6 +18,7 @@ import os
 import re
 import subprocess
 import sys
+from pathlib import Path
 from typing import Callable
 
 
@@ -127,7 +130,7 @@ def versions_from_parentdir(parentdir_prefix, root, verbose):
     rootdirs = []
 
     for _ in range(3):
-        dirname = os.path.basename(root)
+        dirname = Path(root).name
         if dirname.startswith(parentdir_prefix):
             return {
                 "version": dirname[len(parentdir_prefix) :],
@@ -137,7 +140,7 @@ def versions_from_parentdir(parentdir_prefix, root, verbose):
                 "date": None,
             }
         rootdirs.append(root)
-        root = os.path.dirname(root)  # up a level
+        root = Path(root).parent  # up a level
 
     if verbose:
         print(
@@ -358,9 +361,8 @@ def git_pieces_from_vcs(tag_prefix, root, verbose, runner=run_command):
             if verbose:
                 fmt = "tag '%s' doesn't start with prefix '%s'"
                 print(fmt % (full_tag, tag_prefix))
-            pieces["error"] = "tag '{}' doesn't start with prefix '{}'".format(
-                full_tag,
-                tag_prefix,
+            pieces["error"] = (
+                f"tag '{full_tag}' doesn't start with prefix '{tag_prefix}'"
             )
             return pieces
         pieces["closest-tag"] = full_tag[len(tag_prefix) :]
@@ -657,12 +659,12 @@ def get_versions():
         pass
 
     try:
-        root = os.path.realpath(__file__)
+        root = Path(__file__).resolve()
         # versionfile_source is the relative path from the top of the source
         # tree (where the .git directory might live) to this file. Invert
         # this to find the root from __file__.
         for _ in cfg.versionfile_source.split("/"):
-            root = os.path.dirname(root)
+            root = Path(root).parent
     except NameError:
         return {
             "version": "0+unknown",
