@@ -8,61 +8,51 @@ like the ACCESS-NRI catalog
 
 from dataclasses import dataclass
 from functools import partial
-from typing import Callable, Optional
+from typing import Callable
 
 import pandas as pd
 import tlz
 from intake import DataSource
 
 from . import COLUMNS_WITH_ITERABLES
+from .utils import _to_tuple, trace_failure, tuplify_series
+
+__all__ = [
+    "Cmip6Translator",
+    "Cmip5Translator",
+    "BarpaTranslator",
+    "CordexTranslator",
+    "Era5Translator",
+    "EsmValToolTranslator",
+    "CcamTranslator",
+    "NarclimTranslator",
+    "EsgfTranslator",
+]
 
 FREQUENCY_TRANSLATIONS = {
     "monthly-averaged-by-hour": "1hr",
-    "monthly-averaged-by-day": "1hr",
+    "monthly-averaged-by-day": "1day",
+    "1hrCM": "1hr",
     "3hrPt": "3hr",
     "6hrPt": "6hr",
     "daily": "1day",
     "day": "1day",
+    "night": "1day",
+    "mon_day_time": "1mon",
+    "mon_night_time": "1mon",
     "mon": "1mon",
     "monthly-averaged": "1mon",
     "monC": "1mon",
     "monClim": "1mon",
     "monPt": "1mon",
     "sem": "3mon",
+    "20min": "subhr",
     "subhrPt": "subhr",
     "yr": "1yr",
     "yrPt": "1yr",
+    "yrC": "1yr",
+    "na": "fx",
 }
-
-
-def _to_tuple(series: pd.Series) -> pd.Series:
-    """
-    Make each entry in the provided series a tuple
-
-    Parameters
-    ----------
-    series: :py:class:`~pandas.Series`
-        A pandas Series or another object with an `apply` method
-    """
-    return series.apply(lambda x: (x,))
-
-
-def tuplify_series(func: Callable) -> Callable:
-    """
-    Decorator that wraps a function that returns a pandas Series and converts
-    each entry in the series to a tuple
-    """
-
-    def wrapper(*args, **kwargs):
-        # Check if the first argument is 'self'
-        if len(args) > 0 and hasattr(args[0], "__class__"):
-            self = args[0]
-            series = func(self, *args[1:], **kwargs)
-        else:
-            series = func(*args, **kwargs)
-        return _to_tuple(series)
-
-    return wrapper
 
 
 class TranslatorError(Exception):
@@ -149,7 +139,7 @@ class DefaultTranslator:
 
         return pd.Series([val] * len_df)
 
-    def translate(self, groupby: Optional[list[str]] = None) -> pd.DataFrame:
+    def translate(self, groupby: list[str] | None = None) -> pd.DataFrame:
         """
         Return the translated :py:class:`~pandas.DataFrame` of metadata and merge into set of
         set of rows with unique values of the columns specified.
@@ -197,7 +187,7 @@ class DefaultTranslator:
         return df[self.columns]  # Preserve ordering
 
     def set_dispatch(
-        self, core_colname: str, func: Callable, input_name: Optional[str] = None
+        self, core_colname: str, func: Callable, input_name: str | None = None
     ):
         """
         Set a dispatch function for a column. Typically only required when either:
@@ -221,12 +211,14 @@ class DefaultTranslator:
         self._dispatch[core_colname] = func
         setattr(self._dispatch_keys, core_colname, input_name)
 
+    @trace_failure
     def _realm_translator(self) -> pd.Series:
         """
         Return realm, fixing a few issues
         """
         return _cmip_realm_translator(self.source.df[self._dispatch_keys.realm])
 
+    @trace_failure
     @tuplify_series
     def _model_translator(self) -> pd.Series:
         """
@@ -234,6 +226,7 @@ class DefaultTranslator:
         """
         return self.source.df[self._dispatch_keys.model]
 
+    @trace_failure
     @tuplify_series
     def _frequency_translator(self) -> pd.Series:
         """
@@ -243,6 +236,7 @@ class DefaultTranslator:
             lambda x: FREQUENCY_TRANSLATIONS.get(x, x)
         )
 
+    @trace_failure
     @tuplify_series
     def _variable_translator(self) -> pd.Series:
         """
@@ -384,7 +378,9 @@ class CordexTranslator(DefaultTranslator):
 
         super().__init__(source, columns)
         self.set_dispatch(
-            input_name="source_id", core_colname="model", func=super()._model_translator
+            input_name="project_id",
+            core_colname="model",
+            func=super()._model_translator,
         )
         self.set_dispatch(
             input_name="variable_id",
@@ -393,6 +389,11 @@ class CordexTranslator(DefaultTranslator):
         )
         self.set_dispatch(
             input_name="realm", core_colname="realm", func=self._realm_translator
+        )
+        self.set_dispatch(
+            input_name="frequency",
+            core_colname="frequency",
+            func=super()._frequency_translator,
         )
 
     def _realm_translator(self):
@@ -436,6 +437,7 @@ class Era5Translator(DefaultTranslator):
         )
 
     @tuplify_series
+    @trace_failure
     def _model_translator(self):
         """
         Get the model from the path. This is a slightly hacky approach, using the
@@ -444,7 +446,7 @@ class Era5Translator(DefaultTranslator):
         where model is one of 'era5', 'era5t', 'era5-preliminary', 'era5-1',
         'era5-derived'.
         """
-        return self.source.df["path"].str.split("/").str[4]
+        return self.source.df["path"].astype(str).str.split("/").str[4]
 
     def _realm_translator(self):
         """
@@ -454,11 +456,12 @@ class Era5Translator(DefaultTranslator):
         return self.source.df.apply(lambda x: ("none",), 1)
 
     @tuplify_series
+    @trace_failure
     def _frequency_translator(self):
         """
         Get the frequency from the path
         """
-        config_str = self.source.df["path"].str.split("/").str[6].copy()
+        config_str = self.source.df["path"].astype(str).str.split("/").str[6].copy()
         """
         ERA5 contains some datasets where the frequency isn't readily identifiable:
         - 'reanalysis' is at 1hour frequency
@@ -572,16 +575,144 @@ class NarclimTranslator(DefaultTranslator):
         return self.source.df.apply(lambda x: ("atmos",), 1)
 
 
+class EsgfTranslator(DefaultTranslator):
+    def __init__(self, source, columns):
+        """
+        Initialise an EsgfTranslator
+
+        Parameters
+        ----------
+        source: :py:class:`~intake.DataSource`
+            The NCI Earth System Grid intake-esm datastore
+        columns: list of str
+            The columns to translate to (these are the core columns in the intake-dataframe-catalog)
+        """
+
+        super().__init__(source, columns)
+        self.set_dispatch(
+            input_name="source_id",
+            core_colname="model",
+            func=super()._model_translator,
+        )
+        self.set_dispatch(
+            input_name="realm",
+            core_colname="realm",
+            func=super()._realm_translator,
+        )
+        self.set_dispatch(
+            input_name="frequency",
+            core_colname="frequency",
+            func=super()._frequency_translator,
+        )
+        self.set_dispatch(
+            input_name="variable_id",
+            core_colname="variable",
+            func=super()._variable_translator,
+        )
+
+
+class EsmValToolTranslator(DefaultTranslator):
+    """
+    EsmValTool Translator for translating metadata from the NCI EsmValTool intake datastores.
+    """
+
+    def __init__(self, source, columns):
+        """
+        Initialise an EsgfTranslator
+
+        Parameters
+        ----------
+        source: :py:class:`~intake.DataSource`
+            The NCI Earth System Grid intake-esm datastore
+        columns: list of str
+            The columns to translate to (these are the core columns in the intake-dataframe-catalog)
+        """
+
+        super().__init__(source, columns)
+        self.set_dispatch(
+            input_name="source_id",
+            core_colname="model",
+            func=super()._model_translator,
+        )
+        self.set_dispatch(
+            input_name="table_id",
+            core_colname="realm",
+            func=self._realm_translator,
+        )
+        self.set_dispatch(
+            input_name="table_id",
+            core_colname="frequency",
+            func=self._frequency_translator,
+        )
+        self.set_dispatch(
+            input_name="variable_id",
+            core_colname="variable",
+            func=super()._variable_translator,
+        )
+
+    @tuplify_series
+    @trace_failure
+    def _realm_translator(self):
+        """
+        Return realm, fixing a few issues
+        """
+        CT11_TABLEID_REALM_TRANSLATIONS = {
+            "Ofx": "ocean",
+            "OImon": "ocean",
+            "Oyr": "ocean",
+            "Lmon": "land",
+            "aero": "aerosol",
+            "LImon": "landIce",
+            "Amon": "atmos",
+            "Omon": "ocean",
+        }
+
+        return self.source.df["table_id"].apply(
+            lambda x: CT11_TABLEID_REALM_TRANSLATIONS.get(x, "none")
+        )
+
+    @tuplify_series
+    @trace_failure
+    def _frequency_translator(self):
+        """
+        Return frequency, fixing a few issues
+        """
+
+        CT11_TABLEID_FREQ_TRANSLATIONS = {
+            "1.1": "fx",
+            "Ofx": "fx",
+            "OImon": "1mon",
+            "Oyr": "1yr",
+            "day": "1day",
+            "fx": "fx",
+            "Lmon": "1mon",
+            "aero": "fx",
+            "LImon": "1mon",
+            "Amon": "1mon",
+            "mon": "1mon",
+            "Omon": "1mon",
+            "cfMon": "1mon",
+            "Emon": "1mon",
+            "unknown": "fx",
+            "CFday": "1day",
+            "Eday": "1day",
+        }
+
+        return self.source.df["table_id"].apply(
+            lambda x: CT11_TABLEID_FREQ_TRANSLATIONS.get(x, "none")
+        )
+
+
 @dataclass
 class _DispatchKeys:
     """
     Data class to store the keys for the dispatch dictionary in the Translator classes
     """
 
-    model: Optional[str] = None
-    realm: Optional[str] = None
-    frequency: Optional[str] = None
-    variable: Optional[str] = None
+    model: str | None = None
+    realm: str | None = None
+    frequency: str | None = None
+    variable: str | None = None
 
 
 def _cmip_realm_translator(series) -> pd.Series:
