@@ -1,4 +1,5 @@
 # Version: 0.28
+# ruff: noqa: UP031
 
 """The Versioneer - like a rocketeer, but for versions.
 
@@ -1353,9 +1354,8 @@ def git_pieces_from_vcs(tag_prefix, root, verbose, runner=run_command):
             if verbose:
                 fmt = "tag '%s' doesn't start with prefix '%s'"
                 print(fmt % (full_tag, tag_prefix))
-            pieces["error"] = "tag '{}' doesn't start with prefix '{}'".format(
-                full_tag,
-                tag_prefix,
+            pieces["error"] = (
+                f"tag '{full_tag}' doesn't start with prefix '{tag_prefix}'"
             )
             return pieces
         pieces["closest-tag"] = full_tag[len(tag_prefix) :]
