@@ -8,4 +8,4 @@ from . import _version
 __version__ = _version.get_versions()["version"]
 
 CATALOG_LOCATION = "/g/data/xp65/public/apps/access-nri-intake-catalog/catalog.yaml"
-USER_CATALOG_LOCATION = Path.home() / ".access_nri_intake_catalog/catalog.yaml"
+USER_CATALOG_LOCATION = str(Path.home() / ".access_nri_intake_catalog/catalog.yaml")
