@@ -2,16 +2,18 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from pathlib import Path
+from unittest import mock
 
 import intake
 import pandas as pd
 import pytest
 import xarray as xr
+from ecgtools.builder import INVALID_ASSET, TRACEBACK
 from intake_esm.source import _get_xarray_open_kwargs, _open_dataset
 from intake_esm.utils import OPTIONS
 
 from access_nri_intake.source import CORE_COLUMNS, builders
-from access_nri_intake.source.utils import _AccessNCFileInfo
+from access_nri_intake.source.utils import _NCFileInfo
 
 
 @pytest.mark.parametrize(
@@ -28,6 +30,9 @@ from access_nri_intake.source.utils import _AccessNCFileInfo
         ),
         (["access-esm1-5"], "AccessEsm15Builder", {"ensemble": False}, 11, 11, 11),
         (["access-om3"], "AccessOm3Builder", {}, 12, 12, 6),
+        (["mom6"], "Mom6Builder", {}, 27, 27, 15),
+        (["roms"], "ROMSBuilder", {}, 4, 4, 1),
+        (["woa"], "WoaBuilder", {}, 4, 4, 2),
         (["access-mopper"], "MopperBuilder", {"ensemble": True}, 2, 2, 1),
     ],
 )
@@ -133,6 +138,98 @@ def test_builder_build(
             None,
             "GMOM_JRA_WD_ww3_hi_XXXX_XX_XX_XXXXX",
         ),
+        # MOM6
+        (
+            "mom6/output000/19000101.ice_daily.nc",
+            "Mom6Builder",
+            "seaIce",
+            None,
+            "XXXXXXXX_ice_daily",
+        ),
+        (
+            "mom6/output001/19010101.ice_month.nc",
+            "Mom6Builder",
+            "seaIce",
+            None,
+            "XXXXXXXX_ice_month",
+        ),
+        (
+            "mom6/output001/19010101.ocean_annual_rho2.nc",
+            "Mom6Builder",
+            "ocean",
+            None,
+            "XXXXXXXX_ocean_annual_rho2",
+        ),
+        (
+            "mom6/output000/19000101.ocean_annual_z.nc",
+            "Mom6Builder",
+            "ocean",
+            None,
+            "XXXXXXXX_ocean_annual_z",
+        ),
+        (
+            "mom6/output000/19000101.ocean_daily.nc",
+            "Mom6Builder",
+            "ocean",
+            None,
+            "XXXXXXXX_ocean_daily",
+        ),
+        (
+            "mom6/output001/19010101.ocean_month_rho2.nc",
+            "Mom6Builder",
+            "ocean",
+            None,
+            "XXXXXXXX_ocean_month_rho2",
+        ),
+        (
+            "mom6/output000/19000101.ocean_month_z.nc",
+            "Mom6Builder",
+            "ocean",
+            None,
+            "XXXXXXXX_ocean_month_z",
+        ),
+        (
+            "mom6/output001/19010101.ocean_month.nc",
+            "Mom6Builder",
+            "ocean",
+            None,
+            "XXXXXXXX_ocean_month",
+        ),
+        (
+            "mom6/output000/19000101.ocean_scalar_annual.nc",
+            "Mom6Builder",
+            "ocean",
+            None,
+            "XXXXXXXX_ocean_scalar_annual",
+        ),
+        (
+            "mom6/output001/19010101.ocean_scalar_month.nc",
+            "Mom6Builder",
+            "ocean",
+            None,
+            "XXXXXXXX_ocean_scalar_month",
+        ),
+        (
+            "mom6/output000/19000101.ocean_static.nc",
+            "Mom6Builder",
+            "ocean",
+            None,
+            "XXXXXXXX_ocean_static",
+        ),
+        (
+            "roms/roms_his_0016.nc",
+            "ROMSBuilder",
+            "seaIce",
+            None,
+            "roms_his_XXXX",
+        ),
+        (
+            "woa/woa13_ts_01_mom01.nc",
+            "WoaBuilder",
+            "ocean",
+            None,
+            "woa13_ts_XX_mom01",
+        ),
     ],
 )
 def test_builder_parser(test_data, filename, builder, realm, member, file_id):
@@ -142,6 +239,57 @@ def test_builder_parser(test_data, filename, builder, realm, member, file_id):
     if member:
         assert info["member"] == member
     assert info["file_id"] == file_id
+
+
+@mock.patch("access_nri_intake.source.utils._NCFileInfo.to_dict")
+@pytest.mark.parametrize(
+    "filename",
+    [
+        "mom6/output000/19000101.ice_daily.nc",
+        "mom6/output001/19010101.ice_month.nc",
+        "mom6/output001/19010101.ocean_annual_rho2.nc",
+        "mom6/output000/19000101.ocean_annual_z.nc",
+        "mom6/output000/19000101.ocean_daily.nc",
+        "mom6/output001/19010101.ocean_month_rho2.nc",
+        "mom6/output000/19000101.ocean_month_z.nc",
+        "mom6/output001/19010101.ocean_month.nc",
+        "mom6/output000/19000101.ocean_scalar_annual.nc",
+        "mom6/output001/19010101.ocean_scalar_month.nc",
+        "mom6/output000/19000101.ocean_static.nc",
+    ],
+)
+def test_Mom6Builder_parser_bad_realm(to_dict_mock, test_data, filename):
+    to_dict_mock.return_value = {
+        "filename": filename.replace("ice", "badrealm").replace("ocean", "badrealm")
+    }
+    info = builders.Mom6Builder.parser(str(test_data / filename))
+    assert INVALID_ASSET in info.keys()
+    assert TRACEBACK in info.keys()
+    assert "ParserError" in info[TRACEBACK]
+
+
+@pytest.mark.parametrize(
+    "filename",
+    ["rubbish_name.nc"],
+)
+@pytest.mark.parametrize(
+    "builder",
+    [
+        "AccessOm2Builder",
+        "AccessOm3Builder",
+        "Mom6Builder",
+        "AccessEsm15Builder",
+        "AccessCm2Builder",
+        "ROMSBuilder",
+        "WoaBuilder",
+    ],
+)
+def test_builder_parser_exception(test_data, filename, builder):
+    Builder = getattr(builders, builder)
+    info = Builder.parser(str(test_data / filename))
+    assert INVALID_ASSET in info.keys()
+    assert info[INVALID_ASSET] == str(test_data / filename)
+    assert TRACEBACK in info.keys()
 
 
 def test_builder_columns_with_iterables(test_data):
@@ -223,6 +371,11 @@ def test_builder_columns_with_iterables(test_data):
             builders.AccessOm2Builder,
             "iceh.057-daily",
             ("iceh_XXX_daily", "057", (1, "day")),
+        ),
+        (
+            builders.AccessOm2Builder,
+            "iceh.1958-02-daily",
+            ("iceh_XXXX_XX_daily", "1958-02", (1, "day")),
         ),
         (
             builders.AccessOm2Builder,
@@ -377,6 +530,15 @@ def test_builder_columns_with_iterables(test_data):
         ),
         (
             builders.AccessOm3Builder,
+            "access-om3.mom6.static",
+            (
+                "access_om3_mom6_static",
+                None,
+                None,
+            ),
+        ),
+        (
+            builders.AccessOm3Builder,
             "access-om3.mom6.3d.uh.1mon.mean.1900",
             (
                 "access_om3_mom6_3d_uh_1mon_mean_XXXX",
@@ -429,10 +591,218 @@ def test_builder_columns_with_iterables(test_data):
                 None,
             ),
         ),
+        (
+            builders.AccessOm3Builder,
+            "access-om3.mom6.3d.uh.1mon.mean.1900",
+            (
+                "access_om3_mom6_3d_uh_1mon_mean_XXXX",
+                "1900",
+                (1, "mon"),
+            ),
+        ),
+        (
+            builders.AccessOm3Builder,
+            "access-om3.mom6.3d.uh.1mon.mean.1900-01-01-00000",
+            (
+                "access_om3_mom6_3d_uh_1mon_mean_XXXX_XX_XX_XXXXX",
+                "1900-01-01-00000",
+                (1, "mon"),
+            ),
+        ),
+        (
+            builders.AccessOm3Builder,
+            "access-om3.mom6.3d.uh.1mon.mean.1900-01",
+            (
+                "access_om3_mom6_3d_uh_1mon_mean_XXXX_XX",
+                "1900-01",
+                (1, "mon"),
+            ),
+        ),
+        (
+            builders.AccessOm3Builder,
+            "GMOM_JRA_WD.ww3.hi.1900-01-03-00000",
+            (
+                "GMOM_JRA_WD_ww3_hi_XXXX_XX_XX_XXXXX",
+                "1900-01-03-00000",
+                None,
+            ),
+        ),
+        (
+            builders.AccessOm3Builder,
+            "GMOM_JRA_WD.ww3.hi.1900",
+            (
+                "GMOM_JRA_WD_ww3_hi_XXXX",
+                "1900",
+                None,
+            ),
+        ),
+        (
+            builders.AccessOm3Builder,
+            "GMOM_JRA_WD.ww3.hi.1900-01",
+            (
+                "GMOM_JRA_WD_ww3_hi_XXXX_XX",
+                "1900-01",
+                None,
+            ),
+        ),
+        # MOM6
+        (
+            builders.Mom6Builder,
+            "19000101.ice_daily",
+            (
+                "XXXXXXXX_ice_daily",
+                "19000101",
+                (1, "day"),
+            ),
+        ),
+        (
+            builders.Mom6Builder,
+            "19010101.ice_month",
+            (
+                "XXXXXXXX_ice_month",
+                "19010101",
+                (1, "mon"),
+            ),
+        ),
+        (
+            builders.Mom6Builder,
+            "19010101.ocean_annual_rho2",
+            (
+                "XXXXXXXX_ocean_annual_rho2",
+                "19010101",
+                (1, "yr"),
+            ),
+        ),
+        (
+            builders.Mom6Builder,
+            "19000101.ocean_annual_z",
+            (
+                "XXXXXXXX_ocean_annual_z",
+                "19000101",
+                (1, "yr"),
+            ),
+        ),
+        (
+            builders.Mom6Builder,
+            "19000101.ocean_annual",
+            (
+                "XXXXXXXX_ocean_annual",
+                "19000101",
+                (1, "yr"),
+            ),
+        ),
+        (
+            builders.Mom6Builder,
+            "19000101.ocean_daily",
+            (
+                "XXXXXXXX_ocean_daily",
+                "19000101",
+                (1, "day"),
+            ),
+        ),
+        (
+            builders.Mom6Builder,
+            "19010101.ocean_month_rho2",
+            (
+                "XXXXXXXX_ocean_month_rho2",
+                "19010101",
+                (1, "mon"),
+            ),
+        ),
+        (
+            builders.Mom6Builder,
+            "19000101.ocean_month_z",
+            (
+                "XXXXXXXX_ocean_month_z",
+                "19000101",
+                (1, "mon"),
+            ),
+        ),
+        (
+            builders.Mom6Builder,
+            "19000101.ocean_month",
+            (
+                "XXXXXXXX_ocean_month",
+                "19000101",
+                (1, "mon"),
+            ),
+        ),
+        (
+            builders.Mom6Builder,
+            "19000101.ocean_scalar_annual",
+            (
+                "XXXXXXXX_ocean_scalar_annual",
+                "19000101",
+                (1, "yr"),
+            ),
+        ),
+        (
+            builders.Mom6Builder,
+            "19010101.ocean_scalar_month",
+            (
+                "XXXXXXXX_ocean_scalar_month",
+                "19010101",
+                (1, "mon"),
+            ),
+        ),
+        (
+            builders.Mom6Builder,
+            "19010101.ocean_static",
+            (
+                "XXXXXXXX_ocean_static",
+                "19010101",
+                None,
+            ),
+        ),
+        (
+            builders.Mom6Builder,
+            "20000201.ocean_daily_2000_032",
+            (
+                "XXXXXXXX_ocean_daily_XXXX_XXX",
+                "20000201",
+                (1, "day"),
+            ),
+        ),
+        (
+            builders.Mom6Builder,
+            "20000201.ocean_daily_rho2_2000_056",
+            (
+                "XXXXXXXX_ocean_daily_rho2_XXXX_XXX",
+                "20000201",
+                (1, "day"),
+            ),
+        ),
+        (
+            builders.Mom6Builder,
+            "20000201.ocean_daily_z_2000_119",
+            (
+                "XXXXXXXX_ocean_daily_z_XXXX_XXX",
+                "20000201",
+                (1, "day"),
+            ),
+        ),
+        (
+            builders.ROMSBuilder,
+            "roms_his_0016",
+            (
+                "roms_his_XXXX",
+                "0016",
+                None,
+            ),
+        ),
+        (
+            builders.WoaBuilder,
+            "woa13_ts_01_mom01",
+            (
+                "woa13_ts_XX_mom01",
+                "01",
+                None,
+            ),
+        ),
     ],
 )
-def test_parse_access_filename(builder, filename, expected):
-    assert builder.parse_access_filename(filename) == expected
+def test_parse_filename(builder, filename, expected):
+    assert builder.parse_filename(filename) == expected
 
 
 @pytest.mark.parametrize(
@@ -448,7 +818,7 @@ def test_parse_access_filename(builder, filename, expected):
         (
             builders.AccessOm2Builder,
             "access-om2/output000/ocean/ocean_grid.nc",
-            _AccessNCFileInfo(
+            _NCFileInfo(
                 path=None,  # type: ignore
                 filename="ocean_grid.nc",
                 file_id="ocean_grid",
@@ -471,7 +841,7 @@ def test_parse_access_filename(builder, filename, expected):
         (
             builders.AccessOm2Builder,
             "access-om2/output000/ocean/ocean.nc",
-            _AccessNCFileInfo(
+            _NCFileInfo(
                 path=None,  # type: ignore
                 filename="ocean.nc",
                 file_id="ocean",
@@ -521,7 +891,7 @@ def test_parse_access_filename(builder, filename, expected):
         (
             builders.AccessOm2Builder,
             "access-om2/output000/ocean/ocean_month.nc",
-            _AccessNCFileInfo(
+            _NCFileInfo(
                 path=None,  # type: ignore
                 filename="ocean_month.nc",
                 file_id="ocean_month",
@@ -560,7 +930,7 @@ def test_parse_access_filename(builder, filename, expected):
         (
             builders.AccessOm2Builder,
             "access-om2/output000/ocean/ocean_month_inst_nobounds.nc",
-            _AccessNCFileInfo(
+            _NCFileInfo(
                 path=None,  # type: ignore
                 filename="ocean_month_inst_nobounds.nc",
                 file_id="ocean_month_inst_nobounds",
@@ -593,7 +963,7 @@ def test_parse_access_filename(builder, filename, expected):
         (
             builders.AccessOm2Builder,
             "access-om2/output000/ice/OUTPUT/iceh.1900-01.nc",
-            _AccessNCFileInfo(
+            _NCFileInfo(
                 path=None,  # type: ignore
                 filename="iceh.1900-01.nc",
                 file_id="iceh_XXXX_XX",
@@ -625,7 +995,7 @@ def test_parse_access_filename(builder, filename, expected):
         (
             builders.AccessCm2Builder,
             "access-cm2/by578/history/atm/netCDF/by578a.pd201501_dai.nc",
-            _AccessNCFileInfo(
+            _NCFileInfo(
                 path=None,  # type: ignore
                 filename="by578a.pd201501_dai.nc",
                 file_id="by578a_pdXXXXXX_dai",
@@ -643,7 +1013,7 @@ def test_parse_access_filename(builder, filename, expected):
         (
             builders.AccessCm2Builder,
             "access-cm2/by578/history/ice/iceh_d.2015-01.nc",
-            _AccessNCFileInfo(
+            _NCFileInfo(
                 path=None,  # type: ignore
                 filename="iceh_d.2015-01.nc",
                 file_id="iceh_d_XXXX_XX",
@@ -675,7 +1045,7 @@ def test_parse_access_filename(builder, filename, expected):
         (
             builders.AccessCm2Builder,
             "access-cm2/by578/history/ocn/ocean_daily.nc-20150630",
-            _AccessNCFileInfo(
+            _NCFileInfo(
                 path=None,  # type: ignore
                 filename="ocean_daily.nc-20150630",
                 file_id="ocean_daily",
@@ -707,7 +1077,7 @@ def test_parse_access_filename(builder, filename, expected):
         (
             builders.AccessCm2Builder,
             "access-cm2/by578/history/ocn/ocean_scalar.nc-20150630",
-            _AccessNCFileInfo(
+            _NCFileInfo(
                 path=None,  # type: ignore
                 filename="ocean_scalar.nc-20150630",
                 file_id="ocean_scalar",
@@ -749,7 +1119,7 @@ def test_parse_access_filename(builder, filename, expected):
         (
             builders.AccessEsm15Builder,
             "access-esm1-5/history/atm/netCDF/HI-C-05-r1.pa-185001_mon.nc",
-            _AccessNCFileInfo(
+            _NCFileInfo(
                 path=None,  # type: ignore
                 filename="HI-C-05-r1.pa-185001_mon.nc",
                 file_id="HI_C_05_r1_pa_XXXXXX_mon",
@@ -767,7 +1137,7 @@ def test_parse_access_filename(builder, filename, expected):
         (
             builders.AccessEsm15Builder,
             "access-esm1-5/history/ice/iceh.1850-01.nc",
-            _AccessNCFileInfo(
+            _NCFileInfo(
                 path=None,  # type: ignore
                 filename="iceh.1850-01.nc",
                 file_id="iceh_XXXX_XX",
@@ -799,7 +1169,7 @@ def test_parse_access_filename(builder, filename, expected):
         (
             builders.AccessEsm15Builder,
             "access-esm1-5/history/ocn/ocean_bgc_ann.nc-18501231",
-            _AccessNCFileInfo(
+            _NCFileInfo(
                 path=None,  # type: ignore
                 filename="ocean_bgc_ann.nc-18501231",
                 file_id="ocean_bgc_ann",
@@ -838,7 +1208,7 @@ def test_parse_access_filename(builder, filename, expected):
         (
             builders.AccessEsm15Builder,
             "access-esm1-5/history/ocn/ocean_bgc.nc-18501231",
-            _AccessNCFileInfo(
+            _NCFileInfo(
                 path=None,  # type: ignore
                 filename="ocean_bgc.nc-18501231",
                 file_id="ocean_bgc",
@@ -880,7 +1250,7 @@ def test_parse_access_filename(builder, filename, expected):
         (
             builders.AccessOm3Builder,
             "access-om3/output000/GMOM_JRA_WD.mom6.h.native_1900_01.nc",
-            _AccessNCFileInfo(
+            _NCFileInfo(
                 path=None,  # type: ignore
                 filename="GMOM_JRA_WD.mom6.h.native_1900_01.nc",
                 file_id="GMOM_JRA_WD_mom6_h_native_XXXX_XX",
@@ -953,7 +1323,7 @@ def test_parse_access_filename(builder, filename, expected):
         (
             builders.AccessOm3Builder,
             "access-om3/output000/GMOM_JRA_WD.mom6.h.sfc_1900_01_02.nc",
-            _AccessNCFileInfo(
+            _NCFileInfo(
                 path=None,  # type: ignore
                 filename="GMOM_JRA_WD.mom6.h.sfc_1900_01_02.nc",
                 file_id="GMOM_JRA_WD_mom6_h_sfc_XXXX_XX_XX",
@@ -1021,7 +1391,7 @@ def test_parse_access_filename(builder, filename, expected):
         (
             builders.AccessOm3Builder,
             "access-om3/output000/GMOM_JRA_WD.mom6.h.static.nc",
-            _AccessNCFileInfo(
+            _NCFileInfo(
                 path=None,  # type: ignore
                 filename="GMOM_JRA_WD.mom6.h.static.nc",
                 file_id="GMOM_JRA_WD_mom6_h_static",
@@ -1049,7 +1419,7 @@ def test_parse_access_filename(builder, filename, expected):
         (
             builders.AccessOm3Builder,
             "access-om3/output000/GMOM_JRA_WD.mom6.h.z_1900_01.nc",
-            _AccessNCFileInfo(
+            _NCFileInfo(
                 path=None,  # type: ignore
                 filename="GMOM_JRA_WD.mom6.h.z_1900_01.nc",
                 file_id="GMOM_JRA_WD_mom6_h_z_XXXX_XX",
@@ -1122,7 +1492,7 @@ def test_parse_access_filename(builder, filename, expected):
         (
             builders.AccessOm3Builder,
             "access-om3/output000/GMOM_JRA_WD.cice.h.1900-01-01.nc",
-            _AccessNCFileInfo(
+            _NCFileInfo(
                 path=None,  # type: ignore
                 filename="GMOM_JRA_WD.cice.h.1900-01-01.nc",
                 file_id="GMOM_JRA_WD_cice_h_XXXX_XX_XX",
@@ -1154,7 +1524,7 @@ def test_parse_access_filename(builder, filename, expected):
         (
             builders.AccessOm3Builder,
             "access-om3/output000/GMOM_JRA_WD.ww3.hi.1900-01-02-00000.nc",
-            _AccessNCFileInfo(
+            _NCFileInfo(
                 path=None,  # type: ignore
                 filename="GMOM_JRA_WD.ww3.hi.1900-01-02-00000.nc",
                 file_id="GMOM_JRA_WD_ww3_hi_XXXX_XX_XX_XXXXX",
@@ -1169,15 +1539,1488 @@ def test_parse_access_filename(builder, filename, expected):
                 variable_units=["m2 s", "unitless"],
             ),
         ),
+        (
+            builders.Mom6Builder,
+            "mom6/output000/19000101.ice_daily.nc",
+            _NCFileInfo(
+                path=None,  # type: ignore
+                filename="19000101.ice_daily.nc",
+                file_id="XXXXXXXX_ice_daily",
+                filename_timestamp="19000101",
+                frequency="1day",
+                start_date="1900-01-01, 00:00:00",
+                end_date="1900-01-02, 00:00:00",
+                variable=[
+                    "xT",
+                    "xTe",
+                    "yT",
+                    "yTe",
+                    "time",
+                    "nv",
+                    "siconc",
+                    "sithick",
+                    "average_T1",
+                    "average_T2",
+                    "average_DT",
+                    "time_bnds",
+                ],
+                variable_long_name=[
+                    "T point nominal longitude",
+                    "T-cell edge nominal longitude",
+                    "T point nominal latitude",
+                    "T-cell edge nominal latitude",
+                    "time",
+                    "vertex number",
+                    "ice concentration",
+                    "ice thickness",
+                    "Start time for average period",
+                    "End time for average period",
+                    "Length of average period",
+                    "time axis boundaries",
+                ],
+                variable_standard_name=[
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                ],
+                variable_cell_methods=[
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "time: mean",
+                    "time: mean",
+                    "",
+                    "",
+                    "",
+                    "",
+                ],
+                variable_units=[
+                    "degrees_E",
+                    "degrees_E",
+                    "degrees_N",
+                    "degrees_N",
+                    "days since 1900-01-01",
+                    "",
+                    "0-1",
+                    "m-ice",
+                    "days since 1900-01-01",
+                    "days since 1900-01-01",
+                    "days",
+                    "days",
+                ],
+            ),
+        ),
+        (
+            builders.Mom6Builder,
+            "mom6/output000/19000101.ocean_annual_z.nc",
+            _NCFileInfo(
+                path=None,  # type: ignore
+                filename="19000101.ocean_annual_z.nc",
+                file_id="XXXXXXXX_ocean_annual_z",
+                filename_timestamp="19000101",
+                frequency="1yr",
+                start_date="1900-01-01, 00:00:00",
+                end_date="1901-01-01, 00:00:00",
+                variable=[
+                    "xh",
+                    "yh",
+                    "z_l",
+                    "z_i",
+                    "time",
+                    "nv",
+                    "xq",
+                    "yq",
+                    "volcello",
+                    "thetao",
+                    "thetao_xyave",
+                    "so",
+                    "so_xyave",
+                    "agessc",
+                    "uo",
+                    "vo",
+                    "umo",
+                    "vmo",
+                    "uh",
+                    "vh",
+                    "T_adx",
+                    "T_ady",
+                    "S_adx",
+                    "S_ady",
+                    "average_T1",
+                    "average_T2",
+                    "average_DT",
+                    "time_bnds",
+                ],
+                variable_long_name=[
+                    "h point nominal longitude",
+                    "h point nominal latitude",
+                    "Depth at cell center",
+                    "Depth at interface",
+                    "time",
+                    "vertex number",
+                    "q point nominal longitude",
+                    "q point nominal latitude",
+                    "Ocean grid-cell volume",
+                    "Sea Water Potential Temperature",
+                    "Sea Water Potential Temperature",
+                    "Sea Water Salinity",
+                    "Sea Water Salinity",
+                    "Ideal Age Tracer",
+                    "Sea Water X Velocity",
+                    "Sea Water Y Velocity",
+                    "Ocean Mass X Transport",
+                    "Ocean Mass Y Transport",
+                    "Zonal Thickness Flux",
+                    "Meridional Thickness Flux",
+                    "Advective (by residual mean) Zonal Flux of Heat",
+                    "Advective (by residual mean) Meridional Flux of Heat",
+                    "Advective (by residual mean) Zonal Flux of Salt",
+                    "Advective (by residual mean) Meridional Flux of Salt",
+                    "Start time for average period",
+                    "End time for average period",
+                    "Length of average period",
+                    "time axis boundaries",
+                ],
+                variable_standard_name=[
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "ocean_volume",
+                    "sea_water_potential_temperature",
+                    "sea_water_potential_temperature",
+                    "sea_water_salinity",
+                    "sea_water_salinity",
+                    "ideal_age_tracer",
+                    "sea_water_x_velocity",
+                    "sea_water_y_velocity",
+                    "ocean_mass_x_transport",
+                    "ocean_mass_y_transport",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                ],
+                variable_cell_methods=[
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "area:sum z_l:sum yh:sum xh:sum time: mean",
+                    "area:mean z_l:mean yh:mean xh:mean time: mean",
+                    "z_l:mean time: mean",
+                    "area:mean z_l:mean yh:mean xh:mean time: mean",
+                    "z_l:mean time: mean",
+                    "area:mean z_l:mean yh:mean xh:mean time: mean",
+                    "z_l:mean yh:mean xq:point time: mean",
+                    "z_l:mean yq:point xh:mean time: mean",
+                    "z_l:sum yh:sum xq:point time: mean",
+                    "z_l:sum yq:point xh:sum time: mean",
+                    "z_l:sum yh:sum xq:point time: mean",
+                    "z_l:sum yq:point xh:sum time: mean",
+                    "z_l:sum yh:sum xq:point time: mean",
+                    "z_l:sum yq:point xh:sum time: mean",
+                    "z_l:sum yh:sum xq:point time: mean",
+                    "z_l:sum yq:point xh:sum time: mean",
+                    "",
+                    "",
+                    "",
+                    "",
+                ],
+                variable_units=[
+                    "degrees_east",
+                    "degrees_north",
+                    "meters",
+                    "meters",
+                    "days since 1900-01-01",
+                    "",
+                    "degrees_east",
+                    "degrees_north",
+                    "m3",
+                    "degC",
+                    "degC",
+                    "psu",
+                    "psu",
+                    "yr",
+                    "m s-1",
+                    "m s-1",
+                    "kg s-1",
+                    "kg s-1",
+                    "m3 s-1",
+                    "m3 s-1",
+                    "W",
+                    "W",
+                    "psu m3 s-1",
+                    "psu m3 s-1",
+                    "days since 1900-01-01",
+                    "days since 1900-01-01",
+                    "days",
+                    "days",
+                ],
+            ),
+        ),
+        (
+            builders.Mom6Builder,
+            "mom6/output000/19000101.ocean_month_rho2.nc",
+            _NCFileInfo(
+                path=None,  # type: ignore
+                filename="19000101.ocean_month_rho2.nc",
+                file_id="XXXXXXXX_ocean_month_rho2",
+                filename_timestamp="19000101",
+                frequency="1mon",
+                start_date="1900-01-01, 00:00:00",
+                end_date="1900-02-01, 00:00:00",
+                variable=[
+                    "xh",
+                    "yh",
+                    "rho2_l",
+                    "rho2_i",
+                    "time",
+                    "nv",
+                    "yq",
+                    "volcello",
+                    "thkcello",
+                    "vmo",
+                    "average_T1",
+                    "average_T2",
+                    "average_DT",
+                    "time_bnds",
+                ],
+                variable_long_name=[
+                    "h point nominal longitude",
+                    "h point nominal latitude",
+                    "Target Potential Density at cell center",
+                    "Target Potential Density at interface",
+                    "time",
+                    "vertex number",
+                    "q point nominal latitude",
+                    "Ocean grid-cell volume",
+                    "Cell Thickness",
+                    "Ocean Mass Y Transport",
+                    "Start time for average period",
+                    "End time for average period",
+                    "Length of average period",
+                    "time axis boundaries",
+                ],
+                variable_standard_name=[
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "ocean_volume",
+                    "cell_thickness",
+                    "ocean_mass_y_transport",
+                    "",
+                    "",
+                    "",
+                    "",
+                ],
+                variable_cell_methods=[
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "area:sum rho2_l:sum yh:sum xh:sum time: mean",
+                    "area:mean rho2_l:sum yh:mean xh:mean time: mean",
+                    "rho2_l:sum yq:point xh:sum time: mean",
+                    "",
+                    "",
+                    "",
+                    "",
+                ],
+                variable_units=[
+                    "degrees_east",
+                    "degrees_north",
+                    "kg m-3",
+                    "kg m-3",
+                    "days since 1900-01-01",
+                    "",
+                    "degrees_north",
+                    "m3",
+                    "m",
+                    "kg s-1",
+                    "days since 1900-01-01",
+                    "days since 1900-01-01",
+                    "days",
+                    "days",
+                ],
+            ),
+        ),
+        (
+            builders.Mom6Builder,
+            "mom6/output000/19000101.ocean_scalar_annual.nc",
+            _NCFileInfo(
+                path=None,  # type: ignore
+                filename="19000101.ocean_scalar_annual.nc",
+                file_id="XXXXXXXX_ocean_scalar_annual",
+                filename_timestamp="19000101",
+                frequency="1yr",
+                start_date="1900-01-01, 00:00:00",
+                end_date="1901-01-01, 00:00:00",
+                variable=[
+                    "scalar_axis",
+                    "time",
+                    "nv",
+                    "masso",
+                    "volo",
+                    "thetaoga",
+                    "soga",
+                    "sosga",
+                    "average_T1",
+                    "average_T2",
+                    "average_DT",
+                    "time_bnds",
+                ],
+                variable_long_name=[
+                    "none",
+                    "time",
+                    "vertex number",
+                    "Mass of liquid ocean",
+                    "Total volume of liquid ocean",
+                    "Global Mean Ocean Potential Temperature",
+                    "Global Mean Ocean Salinity",
+                    "Sea Surface Salinity",
+                    "Start time for average period",
+                    "End time for average period",
+                    "Length of average period",
+                    "time axis boundaries",
+                ],
+                variable_standard_name=[
+                    "",
+                    "",
+                    "",
+                    "sea_water_mass",
+                    "sea_water_volume",
+                    "sea_water_potential_temperature",
+                    "sea_water_salinity",
+                    "sea_surface_salinity",
+                    "",
+                    "",
+                    "",
+                    "",
+                ],
+                variable_cell_methods=[
+                    "",
+                    "",
+                    "",
+                    "time: mean",
+                    "time: mean",
+                    "time: mean",
+                    "time: mean",
+                    "time: mean",
+                    "",
+                    "",
+                    "",
+                    "",
+                ],
+                variable_units=[
+                    "",
+                    "days since 1900-01-01",
+                    "",
+                    "kg",
+                    "m3",
+                    "degC",
+                    "psu",
+                    "psu",
+                    "days since 1900-01-01",
+                    "days since 1900-01-01",
+                    "days",
+                    "days",
+                ],
+            ),
+        ),
+        (
+            builders.Mom6Builder,
+            "mom6/output000/19000101.ocean_static.nc",
+            _NCFileInfo(
+                path=None,  # type: ignore
+                filename="19000101.ocean_static.nc",
+                file_id="XXXXXXXX_ocean_static",
+                filename_timestamp="19000101",
+                frequency="fx",
+                start_date="1900-01-01, 00:00:00",
+                end_date="1900-01-01, 00:00:00",
+                variable=[
+                    "xh",
+                    "yh",
+                    "time",
+                    "xq",
+                    "yq",
+                    "areacello",
+                    "deptho",
+                    "hfgeou",
+                    "sftof",
+                    "Coriolis",
+                    "geolon",
+                    "geolat",
+                    "geolon_c",
+                    "geolat_c",
+                    "geolon_u",
+                    "geolat_u",
+                    "geolon_v",
+                    "geolat_v",
+                    "wet",
+                    "wet_c",
+                    "wet_u",
+                    "wet_v",
+                    "dxt",
+                    "dyt",
+                    "dxCu",
+                    "dyCu",
+                    "dxCv",
+                    "dyCv",
+                    "areacello_cu",
+                    "areacello_cv",
+                    "areacello_bu",
+                ],
+                variable_long_name=[
+                    "h point nominal longitude",
+                    "h point nominal latitude",
+                    "time",
+                    "q point nominal longitude",
+                    "q point nominal latitude",
+                    "Ocean Grid-Cell Area",
+                    "Sea Floor Depth",
+                    "Upward geothermal heat flux at sea floor",
+                    "Sea Area Fraction",
+                    "Coriolis parameter at corner (Bu) points",
+                    "Longitude of tracer (T) points",
+                    "Latitude of tracer (T) points",
+                    "Longitude of corner (Bu) points",
+                    "Latitude of corner (Bu) points",
+                    "Longitude of zonal velocity (Cu) points",
+                    "Latitude of zonal velocity (Cu) points",
+                    "Longitude of meridional velocity (Cv) points",
+                    "Latitude of meridional velocity (Cv) points",
+                    "0 if land, 1 if ocean at tracer points",
+                    "0 if land, 1 if ocean at corner (Bu) points",
+                    "0 if land, 1 if ocean at zonal velocity (Cu) points",
+                    "0 if land, 1 if ocean at meridional velocity (Cv) points",
+                    "Delta(x) at thickness/tracer points (meter)",
+                    "Delta(y) at thickness/tracer points (meter)",
+                    "Delta(x) at u points (meter)",
+                    "Delta(y) at u points (meter)",
+                    "Delta(x) at v points (meter)",
+                    "Delta(y) at v points (meter)",
+                    "Ocean Grid-Cell Area",
+                    "Ocean Grid-Cell Area",
+                    "Ocean Grid-Cell Area",
+                ],
+                variable_standard_name=[
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "cell_area",
+                    "sea_floor_depth_below_geoid",
+                    "upward_geothermal_heat_flux_at_sea_floor",
+                    "SeaAreaFraction",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "cell_area",
+                    "cell_area",
+                    "cell_area",
+                ],
+                variable_cell_methods=[
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "area:sum yh:sum xh:sum time: point",
+                    "area:mean yh:mean xh:mean time: point",
+                    "area:mean yh:mean xh:mean time: point",
+                    "area:mean yh:mean xh:mean time: point",
+                    "time: point",
+                    "time: point",
+                    "time: point",
+                    "time: point",
+                    "time: point",
+                    "time: point",
+                    "time: point",
+                    "time: point",
+                    "time: point",
+                    "time: point",
+                    "time: point",
+                    "time: point",
+                    "time: point",
+                    "time: point",
+                    "time: point",
+                    "time: point",
+                    "time: point",
+                    "time: point",
+                    "time: point",
+                    "area:sum yh:sum xq:sum time: point",
+                    "area:sum yq:sum xh:sum time: point",
+                    "area:sum yq:sum xq:sum time: point",
+                ],
+                variable_units=[
+                    "degrees_east",
+                    "degrees_north",
+                    "days since 1900-01-01",
+                    "degrees_east",
+                    "degrees_north",
+                    "m2",
+                    "m",
+                    "W m-2",
+                    "%",
+                    "s-1",
+                    "degrees_east",
+                    "degrees_north",
+                    "degrees_east",
+                    "degrees_north",
+                    "degrees_east",
+                    "degrees_north",
+                    "degrees_east",
+                    "degrees_north",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "m",
+                    "m",
+                    "m",
+                    "m",
+                    "m",
+                    "m",
+                    "m2",
+                    "m2",
+                    "m2",
+                ],
+            ),
+        ),
+        (
+            builders.Mom6Builder,
+            "mom6/output053/20051101.ocean_daily_2005_360.nc",
+            _NCFileInfo(
+                path=None,  # type: ignore
+                filename="20051101.ocean_daily_2005_360.nc",
+                file_id="XXXXXXXX_ocean_daily_XXXX_XXX",
+                filename_timestamp="20051101",
+                frequency="1day",
+                start_date="2005-12-26, 00:00:00",
+                end_date="2005-12-27, 00:00:00",
+                variable=[
+                    "xh",
+                    "yh",
+                    "zl",
+                    "time",
+                    "nv",
+                    "xq",
+                    "yq",
+                    "volcello",
+                    "zos",
+                    "mlotst",
+                    "tauuo",
+                    "tauvo",
+                    "hfds",
+                    "wfo",
+                    "fsitherm",
+                    "salt_flux",
+                    "T_adx_2d",
+                    "T_ady_2d",
+                    "average_T1",
+                    "average_T2",
+                    "average_DT",
+                    "time_bnds",
+                ],
+                variable_long_name=[
+                    "h point nominal longitude",
+                    "h point nominal latitude",
+                    "Layer pseudo-depth, -z*",
+                    "time",
+                    "vertex number",
+                    "q point nominal longitude",
+                    "q point nominal latitude",
+                    "Ocean grid-cell volume",
+                    "Sea surface height above geoid",
+                    "Ocean Mixed Layer Thickness Defined by Sigma T",
+                    "Surface Downward X Stress",
+                    "Surface Downward Y Stress",
+                    "Surface ocean heat flux from SW+LW+latent+sensible+masstransfer+frazil+seaice_melt_heat",
+                    "Water Flux Into Sea Water",
+                    "water flux to ocean from sea ice melt(> 0) or form(< 0)",
+                    "Net salt flux into ocean at surface (restoring + sea-ice)",
+                    "Advective (by residual mean) Zonal Flux of Heat",
+                    "Advective (by residual mean) Meridional Flux of Heat",
+                    "Start time for average period",
+                    "End time for average period",
+                    "Length of average period",
+                    "time axis boundaries",
+                ],
+                variable_standard_name=[
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "ocean_volume",
+                    "sea_surface_height_above_geoid",
+                    "ocean_mixed_layer_thickness_defined_by_sigma_t",
+                    "surface_downward_x_stress",
+                    "surface_downward_y_stress",
+                    "surface_downward_heat_flux_in_sea_water",
+                    "water_flux_into_sea_water",
+                    "water_flux_into_sea_water_due_to_sea_ice_thermodynamics",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                ],
+                variable_cell_methods=[
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "area:sum zl:sum yh:sum xh:sum time: mean",
+                    "area:mean yh:mean xh:mean time: mean",
+                    "area:mean yh:mean xh:mean time: mean",
+                    "yh:mean xq:point time: mean",
+                    "yq:point xh:mean time: mean",
+                    "area:mean yh:mean xh:mean time: mean",
+                    "area:mean yh:mean xh:mean time: mean",
+                    "area:mean yh:mean xh:mean time: mean",
+                    "area:mean yh:mean xh:mean time: mean",
+                    "zl:sum yh:sum xq:point time: mean",
+                    "zl:sum yq:point xh:sum time: mean",
+                    "",
+                    "",
+                    "",
+                    "",
+                ],
+                variable_units=[
+                    "degrees_east",
+                    "degrees_north",
+                    "meter",
+                    "days since 1991-01-01",
+                    "",
+                    "degrees_east",
+                    "degrees_north",
+                    "m3",
+                    "m",
+                    "m",
+                    "N m-2",
+                    "N m-2",
+                    "W m-2",
+                    "kg m-2 s-1",
+                    "kg m-2 s-1",
+                    "kg m-2 s-1",
+                    "W",
+                    "W",
+                    "days since 1991-01-01",
+                    "days since 1991-01-01",
+                    "days",
+                    "days",
+                ],
+            ),
+        ),
+        (
+            builders.Mom6Builder,
+            "mom6/output053/20051101.ocean_daily_rho2_2005_360.nc",
+            _NCFileInfo(
+                path=None,  # type: ignore
+                filename="20051101.ocean_daily_rho2_2005_360.nc",
+                file_id="XXXXXXXX_ocean_daily_rho2_XXXX_XXX",
+                filename_timestamp="20051101",
+                frequency="1day",
+                start_date="2005-12-26, 00:00:00",
+                end_date="2005-12-27, 00:00:00",
+                variable=[
+                    "xh",
+                    "yh",
+                    "rho2_l",
+                    "rho2_i",
+                    "time",
+                    "nv",
+                    "xq",
+                    "yq",
+                    "volcello",
+                    "umo",
+                    "vmo",
+                    "average_T1",
+                    "average_T2",
+                    "average_DT",
+                    "time_bnds",
+                ],
+                variable_long_name=[
+                    "h point nominal longitude",
+                    "h point nominal latitude",
+                    "Target Potential Density at cell center",
+                    "Target Potential Density at interface",
+                    "time",
+                    "vertex number",
+                    "q point nominal longitude",
+                    "q point nominal latitude",
+                    "Ocean grid-cell volume",
+                    "Ocean Mass X Transport",
+                    "Ocean Mass Y Transport",
+                    "Start time for average period",
+                    "End time for average period",
+                    "Length of average period",
+                    "time axis boundaries",
+                ],
+                variable_standard_name=[
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "ocean_volume",
+                    "ocean_mass_x_transport",
+                    "ocean_mass_y_transport",
+                    "",
+                    "",
+                    "",
+                    "",
+                ],
+                variable_cell_methods=[
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "area:sum rho2_l:sum yh:sum xh:sum time: mean",
+                    "rho2_l:sum yh:sum xq:point time: mean",
+                    "rho2_l:sum yq:point xh:sum time: mean",
+                    "",
+                    "",
+                    "",
+                    "",
+                ],
+                variable_units=[
+                    "degrees_east",
+                    "degrees_north",
+                    "kg m-3",
+                    "kg m-3",
+                    "days since 1991-01-01",
+                    "",
+                    "degrees_east",
+                    "degrees_north",
+                    "m3",
+                    "kg s-1",
+                    "kg s-1",
+                    "days since 1991-01-01",
+                    "days since 1991-01-01",
+                    "days",
+                    "days",
+                ],
+            ),
+        ),
+        (
+            builders.Mom6Builder,
+            "mom6/output053/20051101.ocean_daily_z_2005_360.nc",
+            _NCFileInfo(
+                path=None,  # type: ignore
+                filename="20051101.ocean_daily_z_2005_360.nc",
+                file_id="XXXXXXXX_ocean_daily_z_XXXX_XXX",
+                filename_timestamp="20051101",
+                frequency="1day",
+                start_date="2005-12-26, 00:00:00",
+                end_date="2005-12-27, 00:00:00",
+                variable=[
+                    "xh",
+                    "yh",
+                    "z_l",
+                    "z_i",
+                    "time",
+                    "nv",
+                    "xq",
+                    "z_l_sub01",
+                    "yq",
+                    "volcello",
+                    "uo",
+                    "vo",
+                    "thetao",
+                    "so",
+                    "average_T1",
+                    "average_T2",
+                    "average_DT",
+                    "time_bnds",
+                ],
+                variable_long_name=[
+                    "h point nominal longitude",
+                    "h point nominal latitude",
+                    "Depth at cell center",
+                    "Depth at interface",
+                    "time",
+                    "vertex number",
+                    "q point nominal longitude",
+                    "Depth at cell center",
+                    "q point nominal latitude",
+                    "Ocean grid-cell volume",
+                    "Sea Water X Velocity",
+                    "Sea Water Y Velocity",
+                    "Sea Water Potential Temperature",
+                    "Sea Water Salinity",
+                    "Start time for average period",
+                    "End time for average period",
+                    "Length of average period",
+                    "time axis boundaries",
+                ],
+                variable_standard_name=[
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "ocean_volume",
+                    "sea_water_x_velocity",
+                    "sea_water_y_velocity",
+                    "sea_water_potential_temperature",
+                    "sea_water_salinity",
+                    "",
+                    "",
+                    "",
+                    "",
+                ],
+                variable_cell_methods=[
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "area:sum z_l:sum yh:sum xh:sum time: mean",
+                    "z_l:mean yh:mean xq:point time: mean",
+                    "z_l:mean yq:point xh:mean time: mean",
+                    "area:mean z_l:mean yh:mean xh:mean time: mean",
+                    "area:mean z_l:mean yh:mean xh:mean time: mean",
+                    "",
+                    "",
+                    "",
+                    "",
+                ],
+                variable_units=[
+                    "degrees_east",
+                    "degrees_north",
+                    "meters",
+                    "meters",
+                    "days since 1991-01-01",
+                    "",
+                    "degrees_east",
+                    "meters",
+                    "degrees_north",
+                    "m3",
+                    "m s-1",
+                    "m s-1",
+                    "degC",
+                    "psu",
+                    "days since 1991-01-01",
+                    "days since 1991-01-01",
+                    "days",
+                    "days",
+                ],
+            ),
+        ),
+        (
+            builders.ROMSBuilder,
+            "roms/roms_his_0016.nc",
+            _NCFileInfo(
+                filename="roms_his_0016.nc",
+                file_id="roms_his_XXXX",
+                path=None,  # type: ignore
+                filename_timestamp="0016",
+                frequency="fx",
+                start_date="none",
+                end_date="none",
+                variable=[
+                    "ntimes",
+                    "ndtfast",
+                    "dt",
+                    "dtfast",
+                    "dstart",
+                    "nHIS",
+                    "ndefHIS",
+                    "nRST",
+                    "ntsAVG",
+                    "nAVG",
+                    "ndefAVG",
+                    "Falpha",
+                    "Fbeta",
+                    "Fgamma",
+                    "nl_tnu2",
+                    "nl_visc2",
+                    "LuvSponge",
+                    "LtracerSponge",
+                    "Akt_bak",
+                    "Akv_bak",
+                    "rdrg",
+                    "rdrg2",
+                    "Zob",
+                    "Zos",
+                    "Znudg",
+                    "M2nudg",
+                    "M3nudg",
+                    "Tnudg",
+                    "FSobc_in",
+                    "FSobc_out",
+                    "M2obc_in",
+                    "M2obc_out",
+                    "Tobc_in",
+                    "Tobc_out",
+                    "M3obc_in",
+                    "M3obc_out",
+                    "rho0",
+                    "gamma2",
+                    "LuvSrc",
+                    "LwSrc",
+                    "LtracerSrc",
+                    "LsshCLM",
+                    "Lm2CLM",
+                    "Lm3CLM",
+                    "LtracerCLM",
+                    "LnudgeM2CLM",
+                    "LnudgeM3CLM",
+                    "LnudgeTCLM",
+                    "spherical",
+                    "xl",
+                    "el",
+                    "Vtransform",
+                    "Vstretching",
+                    "theta_s",
+                    "theta_b",
+                    "Tcline",
+                    "hc",
+                    "s_rho",
+                    "s_w",
+                    "Cs_r",
+                    "Cs_w",
+                    "h",
+                    "zice",
+                    "f",
+                    "pm",
+                    "pn",
+                    "x_rho",
+                    "y_rho",
+                    "x_u",
+                    "y_u",
+                    "x_v",
+                    "y_v",
+                    "x_psi",
+                    "y_psi",
+                    "angle",
+                    "mask_rho",
+                    "mask_u",
+                    "mask_v",
+                    "mask_psi",
+                    "ocean_time",
+                    "m",
+                    "Tb",
+                    "Sb",
+                    "zeta",
+                    "ubar",
+                    "vbar",
+                    "u",
+                    "v",
+                    "w",
+                    "temp",
+                    "salt",
+                    "rho",
+                    "Hsbl",
+                    "AKv",
+                    "AKt",
+                    "shflux",
+                    "ssflux",
+                    "sustr",
+                    "svstr",
+                ],
+                variable_long_name=[
+                    "number of long time-steps",
+                    "number of short time-steps",
+                    "size of long time-steps",
+                    "size of short time-steps",
+                    "time stamp assigned to model initilization",
+                    "number of time-steps between history records",
+                    "number of time-steps between the creation of history files",
+                    "number of time-steps between restart records",
+                    "starting time-step for accumulation of time-averaged fields",
+                    "number of time-steps between time-averaged records",
+                    "number of time-steps between the creation of average files",
+                    "Power-law shape barotropic filter parameter",
+                    "Power-law shape barotropic filter parameter",
+                    "Power-law shape barotropic filter parameter",
+                    "nonlinear model Laplacian mixing coefficient for tracers",
+                    "nonlinear model Laplacian mixing coefficient for momentum",
+                    "horizontal viscosity sponge activation switch",
+                    "horizontal diffusivity sponge activation switch",
+                    "background vertical mixing coefficient for tracers",
+                    "background vertical mixing coefficient for momentum",
+                    "linear drag coefficient",
+                    "quadratic drag coefficient",
+                    "bottom roughness",
+                    "surface roughness",
+                    "free-surface nudging/relaxation inverse time scale",
+                    "2D momentum nudging/relaxation inverse time scale",
+                    "3D momentum nudging/relaxation inverse time scale",
+                    "Tracers nudging/relaxation inverse time scale",
+                    "free-surface inflow, nudging inverse time scale",
+                    "free-surface outflow, nudging inverse time scale",
+                    "2D momentum inflow, nudging inverse time scale",
+                    "2D momentum outflow, nudging inverse time scale",
+                    "tracers inflow, nudging inverse time scale",
+                    "tracers outflow, nudging inverse time scale",
+                    "3D momentum inflow, nudging inverse time scale",
+                    "3D momentum outflow, nudging inverse time scale",
+                    "mean density used in Boussinesq approximation",
+                    "slipperiness parameter",
+                    "momentum point sources and sink activation switch",
+                    "mass point sources and sink activation switch",
+                    "tracer point sources and sink activation switch",
+                    "sea surface height climatology processing switch",
+                    "2D momentum climatology processing switch",
+                    "3D momentum climatology processing switch",
+                    "tracer climatology processing switch",
+                    "2D momentum climatology nudging activation switch",
+                    "3D momentum climatology nudging activation switch",
+                    "tracer climatology nudging activation switch",
+                    "grid type logical switch",
+                    "domain length in the XI-direction",
+                    "domain length in the ETA-direction",
+                    "vertical terrain-following transformation equation",
+                    "vertical terrain-following stretching function",
+                    "S-coordinate surface control parameter",
+                    "S-coordinate bottom control parameter",
+                    "S-coordinate surface/bottom layer width",
+                    "S-coordinate parameter, critical depth",
+                    "S-coordinate at RHO-points",
+                    "S-coordinate at W-points",
+                    "S-coordinate stretching curves at RHO-points",
+                    "S-coordinate stretching curves at W-points",
+                    "bathymetry at RHO-points",
+                    "ice draft at RHO-points",
+                    "Coriolis parameter at RHO-points",
+                    "curvilinear coordinate metric in XI",
+                    "curvilinear coordinate metric in ETA",
+                    "x-locations of RHO-points",
+                    "y-locations of RHO-points",
+                    "x-locations of U-points",
+                    "y-locations of U-points",
+                    "x-locations of V-points",
+                    "y-locations of V-points",
+                    "x-locations of PSI-points",
+                    "y-locations of PSI-points",
+                    "angle between XI-axis and EAST",
+                    "mask on RHO-points",
+                    "mask on U-points",
+                    "mask on V-points",
+                    "mask on psi-points",
+                    "time since initialization",
+                    "melt rate",
+                    "ice base temperature",
+                    "ice base salinity",
+                    "free-surface",
+                    "vertically integrated u-momentum component",
+                    "vertically integrated v-momentum component",
+                    "u-momentum component",
+                    "v-momentum component",
+                    "vertical momentum component",
+                    "potential temperature",
+                    "salinity",
+                    "density anomaly",
+                    "depth of oceanic surface boundary layer",
+                    "vertical viscosity coefficient",
+                    "temperature vertical diffusion coefficient",
+                    "surface net heat flux",
+                    "kinematic surface net salt flux, SALT (E-P)/rhow",
+                    "surface u-momentum stress",
+                    "surface v-momentum stress",
+                ],
+                variable_standard_name=[
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "ocean_s_coordinate_g2",
+                    "ocean_s_coordinate_g2",
+                    "",
+                    "",
+                    "sea_floor_depth",
+                    "",
+                    "coriolis_parameter",
+                    "inverse_grid_x_spacing",
+                    "inverse_grid_y_spacing",
+                    "grid_x_location_at_cell_center",
+                    "grid_y_location_at_cell_center",
+                    "grid_x_location_at_cell_y_edges",
+                    "grid_y_location_at_cell_y_edges",
+                    "grid_x_location_at_cell_x_edges",
+                    "grid_y_location_at_cell_x_edges",
+                    "grid_x_location_at_cell_corners",
+                    "grid_y_location_at_cell_corners",
+                    "grid_angle_of_rotation_from_east_to_y",
+                    "land_sea_mask_at_cell_center",
+                    "land_sea_mask_at_cell_y_edges",
+                    "land_sea_mask_at_cell_x_edges",
+                    "land_sea_mask_at_cell_corners",
+                    "time",
+                    "ice_shelf_melt_rate",
+                    "temperature_at_ice_base",
+                    "salinity_at_ice_base",
+                    "sea_surface_elevation_anomaly",
+                    "barotropic_sea_water_x_velocity",
+                    "barotropic_sea_water_y_velocity",
+                    "sea_water_x_velocity",
+                    "sea_water_y_velocity",
+                    "upward_sea_water_velocity",
+                    "sea_water_potential_temperature",
+                    "sea_water_practical_salinity",
+                    "sea_water_density_anomaly",
+                    "ocean_surface_boundary_layer_thickness",
+                    "vertical_viscosity_coefficient_of_sea_water",
+                    "vertical_diffusion_coefficient_of_temperature_in_sea_water",
+                    "surface_downward_heat_flux_in_sea_water",
+                    "salt_flux_into_sea_water",
+                    "",
+                    "surface_downward_x_stress",
+                ],
+                variable_cell_methods=[
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "ocean_time: point",
+                    "ocean_time: point",
+                    "ocean_time: point",
+                    "ocean_time: point",
+                    "ocean_time: point",
+                    "ocean_time: point",
+                    "ocean_time: point",
+                    "ocean_time: point",
+                    "ocean_time: point",
+                    "ocean_time: point",
+                    "ocean_time: point",
+                    "ocean_time: point",
+                    "ocean_time: point",
+                    "ocean_time: point",
+                    "ocean_time: point",
+                    "ocean_time: point",
+                    "ocean_time: point",
+                    "ocean_time: point",
+                    "ocean_time: point",
+                ],
+                variable_units=[
+                    "",
+                    "",
+                    "second",
+                    "second",
+                    "days since 2007-01-01 00:00:00",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "meter2 second-1",
+                    "meter2 second-1",
+                    "",
+                    "",
+                    "meter2 second-1",
+                    "meter2 second-1",
+                    "meter second-1",
+                    "",
+                    "meter",
+                    "meter",
+                    "day-1",
+                    "day-1",
+                    "day-1",
+                    "day-1",
+                    "second-1",
+                    "second-1",
+                    "second-1",
+                    "second-1",
+                    "second-1",
+                    "second-1",
+                    "second-1",
+                    "second-1",
+                    "kilogram meter-3",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "meter",
+                    "meter",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "meter",
+                    "meter",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "meter",
+                    "meter",
+                    "second-1",
+                    "meter-1",
+                    "meter-1",
+                    "meter",
+                    "meter",
+                    "meter",
+                    "meter",
+                    "meter",
+                    "meter",
+                    "meter",
+                    "meter",
+                    "radians",
+                    "",
+                    "",
+                    "",
+                    "",
+                    "seconds since 2007-01-01 00:00:00",
+                    "meter second-1",
+                    "Celsius",
+                    "",
+                    "meter",
+                    "meter second-1",
+                    "meter second-1",
+                    "meter second-1",
+                    "meter second-1",
+                    "meter second-1",
+                    "Celsius",
+                    "",
+                    "kilogram meter-3",
+                    "meter",
+                    "meter2 second-1",
+                    "meter2 second-1",
+                    "watt meter-2",
+                    "meter second-1",
+                    "newton meter-2",
+                    "newton meter-2",
+                ],
+            ),
+        ),
     ],
 )
 def test_parse_access_ncfile(test_data, builder, filename, expected, compare_files):
+    """
+    Tests for correctness of parser. Note that if we are using intake-esm without
+    coordinate discovery enabled, the `conftest.py` fixture will dynamically set
+    the `xfails=True` on parametrisations of these tests where we compare files.
+    This is currently default behaviour as the latest release of intake-esm does
+    not support coordinate discovery. However, it is on main.
+    """
     file = str(test_data / Path(filename))
 
     # Set the path to the test data directory
     expected.path = file
 
-    assert builder.parse_access_ncfile(file) == expected
+    parsed_file = builder.parse_ncfile(file)
+
+    exptd_dict = expected.to_dict()
+    parsed_dict = parsed_file.to_dict()
+    assert exptd_dict == parsed_dict
 
     if not compare_files:
         return None
@@ -1207,3 +3050,84 @@ def test_parse_access_ncfile(test_data, builder, filename, expected, compare_fil
     xr_ds = xr_ds[expected.variable]
 
     xr.testing.assert_equal(ie_ds, xr_ds)
+
+
+@pytest.mark.parametrize(
+    "builder,default_include,default_exclude",
+    [
+        (builders.BaseBuilder, [], []),
+        (builders.AccessOm2Builder, ["*.nc"], ["*restart*", "*o2i.nc"]),
+        (
+            builders.AccessOm3Builder,
+            ["*.nc"],
+            [
+                "*restart*",
+                "*MOM_IC.nc",
+                "*ocean_geometry.nc",
+                "*ocean.stats.nc",
+                "*Vertical_coordinate.nc",
+            ],
+        ),
+        (
+            builders.Mom6Builder,
+            ["*.nc"],
+            [
+                "*restart*",
+                "*MOM_IC.nc",
+                "*sea_ice_geometry.nc",
+                "*ocean_geometry.nc",
+                "*ocean.stats.nc",
+                "*Vertical_coordinate.nc",
+            ],
+        ),
+        (builders.AccessEsm15Builder, ["*.nc*"], ["*restart*"]),
+        (builders.AccessCm2Builder, ["*.nc*"], ["*restart*"]),
+        (builders.ROMSBuilder, ["*.nc"], ["*avg*", "*rst*"]),
+    ],
+)
+@pytest.mark.parametrize("include_patts", [None, ["include", "patterns"]])
+@pytest.mark.parametrize("exclude_patts", [None, ["exclude", "patterns"]])
+def test_builder_include_exclude_patterns(
+    builder, default_include, default_exclude, include_patts, exclude_patts, tmpdir
+):
+    if builder == builders.AccessEsm15Builder or builder == builders.AccessCm2Builder:
+        additional = [
+            False,  # ensemble needed for ESM builder
+        ]
+    else:
+        additional = []
+
+    kwargs = {}
+    if include_patts is not None:
+        kwargs["include_patterns"] = include_patts
+    if exclude_patts is not None:
+        kwargs["exclude_patterns"] = exclude_patts
+
+    bld = builder(str(tmpdir), *additional, **kwargs)
+    assert bld.include_patterns == (
+        include_patts if include_patts is not None else default_include
+    )
+    assert bld.exclude_patterns == (
+        exclude_patts if exclude_patts is not None else default_exclude
+    )
+
+
+def test_parser_woa(test_data, tmpdir):
+    """
+    Test the parser for WOA files separates out kds/non kds files correctly.
+
+    This kind of duplicates test_builder_build, but hopefully having it separate
+    keeps the intent clear. When we move to grid based file ids this can be deleted.
+    """
+    srcdir = test_data / "woa"
+
+    builders.WoaBuilder(str(srcdir)).build().save(
+        name="woa_ds", description="test_woa", directory=str(tmpdir)
+    )
+
+    esm_ds = intake.open_esm_datastore(
+        str(tmpdir / "woa_ds.json"),
+    )
+
+    # Should be two datastores.
+    assert len(esm_ds) == 2
