@@ -2,17 +2,18 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import re
+from pathlib import Path
 from unittest import mock
 
 import pytest
 
 import access_nri_intake
 from access_nri_intake.data import CATALOG_NAME_FORMAT
-from access_nri_intake.data.utils import _get_catalog_rp, available_versions
+from access_nri_intake.data.utils import _get_catalog_root, available_versions
 
 
 @mock.patch("access_nri_intake.data.utils.get_catalog_fp")
-def test__get_catalog_rp(mock_get_catalog_fp, test_data):
+def test__get_catalog_root(mock_get_catalog_fp, test_data):
     """
     Check that we correctly decipher to rootpath (rp) to the catalogs
     """
@@ -22,9 +23,9 @@ def test__get_catalog_rp(mock_get_catalog_fp, test_data):
         == test_data / "catalog/catalog-good.yaml"
     ), "Mock failed"
 
-    rp = _get_catalog_rp()
-    assert (
-        rp == "/this/is/root/path/"
+    rp = _get_catalog_root()
+    assert rp == Path(
+        "/this/is/root/path/"
     ), f"Computed root path {rp} != expected value /this/is/root/path/"
 
 
@@ -32,7 +33,7 @@ def test__get_catalog_rp(mock_get_catalog_fp, test_data):
 @pytest.mark.parametrize(
     "cat", ["catalog/catalog-bad-path.yaml", "catalog/catalog-bad-structure.yaml"]
 )
-def test__get_catalog_rp_runtime_errors(mock_get_catalog_fp, test_data, cat):
+def test__get_catalog_root_runtime_errors(mock_get_catalog_fp, test_data, cat):
     """
     Check that we correctly decipher to rootpath (rp) to the catalogs
     """
@@ -42,12 +43,14 @@ def test__get_catalog_rp_runtime_errors(mock_get_catalog_fp, test_data, cat):
     ), "Mock failed"
 
     with pytest.raises(RuntimeError):
-        _get_catalog_rp()
+        _get_catalog_root()
 
 
-@mock.patch("access_nri_intake.data.utils._get_catalog_rp")
-def test_available_versions(mock__get_catalog_rp, test_data):
-    mock__get_catalog_rp.return_value = test_data / "catalog/catalog-dirs"
+@mock.patch("access_nri_intake.data.utils._get_catalog_root")
+@mock.patch("access_nri_intake.data.utils.get_catalog_fp")
+def test_available_versions(mock_get_catalog_fp, mock__get_catalog_root, test_data):
+    mock__get_catalog_root.return_value = test_data / "catalog/catalog-dirs"
+    mock_get_catalog_fp.return_value = test_data / "catalog/catalog-versions.yaml"
     cats = available_versions(pretty=False)
     assert cats == [
         "v2025-02-28",
@@ -57,14 +60,48 @@ def test_available_versions(mock__get_catalog_rp, test_data):
     ], "Did not get expected catalog list"
 
 
-@mock.patch("access_nri_intake.data.utils._get_catalog_rp")
-def test_available_versions_pretty(mock__get_catalog_rp, test_data, capfd):
-    mock__get_catalog_rp.return_value = test_data / "catalog/catalog-dirs"
+@mock.patch("access_nri_intake.data.utils._get_catalog_root")
+@mock.patch("access_nri_intake.data.utils.get_catalog_fp")
+def test_available_versions_pretty(
+    mock_get_catalog_fp, mock__get_catalog_root, test_data, capfd
+):
+    mock__get_catalog_root.return_value = test_data / "catalog/catalog-dirs"
+    mock_get_catalog_fp.return_value = test_data / "catalog/catalog-versions.yaml"
     available_versions(pretty=True)
     captured, _ = capfd.readouterr()
     assert (
-        captured == "v2025-02-28\nv2024-06-19\nv2024-01-01\nv2019-02-02(-->vN.N.N)\n"
+        captured
+        == "v2025-02-28*\nv2024-06-19\nv2024-01-01\nv2019-02-02(-->vN.N.N)\n\nDeprecated catalog catalog-versions-old.yaml:\nv2016-12-31*\nv2016-06-15\nv2016-01-01\n"
     ), "Did not get expected catalog printout"
+
+
+@mock.patch("access_nri_intake.data.utils._get_catalog_root")
+@mock.patch("access_nri_intake.data.utils.get_catalog_fp")
+@mock.patch("pathlib.Path.glob")
+def test_available_versions_pretty_missing_old_catalog_files(
+    mock_glob, mock_get_catalog_fp, mock__get_catalog_root, test_data, capfd
+):
+    mock__get_catalog_root.return_value = test_data / "catalog/catalog-dirs"
+    mock_get_catalog_fp.return_value = test_data / "catalog/catalog-versions.yaml"
+    mock_glob.return_value = [Path("nope.yaml"), Path("nope2.yaml")]
+
+    with pytest.warns(UserWarning, match="Unable to find old catalog file"):
+        available_versions(pretty=True)
+
+
+@mock.patch(
+    "access_nri_intake.data.utils.get_catalog_fp", return_value="/this/is/not/real.yaml"
+)
+def test_available_versions_no_catalog(mock_get_catalog_fp):
+    with pytest.raises(FileNotFoundError):
+        available_versions()
+
+
+@mock.patch("access_nri_intake.data.utils.get_catalog_fp")
+def test_available_versions_bad_catalog(mock_get_catalog_fp, test_data):
+    mock_get_catalog_fp.return_value = test_data / "catalog/catalog-bad-structure.yaml"
+    with pytest.raises(RuntimeError):
+        available_versions()
 
 
 @pytest.mark.parametrize(
