@@ -15,10 +15,13 @@ from access_nri_intake.catalog.translators import (
     CordexTranslator,
     DefaultTranslator,
     Era5Translator,
+    EsgfTranslator,
+    EsmValToolTranslator,
     NarclimTranslator,
     TranslatorError,
     _cmip_realm_translator,
     _to_tuple,
+    trace_failure,
     tuplify_series,
 )
 
@@ -141,7 +144,7 @@ def test_cmip_realm_translator(input, expected):
 def test_to_tuple(input):
     """Test the _to_tuple function"""
     series = pd.Series(input)
-    assert all(_to_tuple(series).map(type) == tuple)
+    assert all(_to_tuple(series).map(type) == tuple)  # noqa: E721
 
 
 @pytest.mark.parametrize(
@@ -181,7 +184,7 @@ def test_DefaultTranslator(test_data, name, description, something):
     assert all(df["name"].to_numpy() == name)
     assert all(df["description"].to_numpy() == description)
     assert all(df["something"].to_numpy() == something)
-    assert all(df["variable"].map(type) == tuple)
+    assert all(df["variable"].map(type) == tuple)  # noqa: E721
     assert all(df["version"].str.startswith("v"))
 
     if name:
@@ -361,3 +364,83 @@ def test_NarclimTranslator(test_data, groupby, n_entries):
     esmds.description = "description"
     df = NarclimTranslator(esmds, CORE_COLUMNS).translate(groupby)
     assert len(df) == n_entries
+
+
+@pytest.mark.parametrize(
+    "groupby, n_entries",
+    [
+        (None, 6),
+        (["variable"], 4),
+        (
+            ["frequency"],
+            2,
+        ),  # We have 3hr and 3hrPt in the test dataset - both translate to 3hr.
+        (["model"], 2),
+        (["realm"], 1),
+    ],
+)
+def test_esgfTranslator(test_data, groupby, n_entries):
+    esmds = intake.open_esm_datastore(
+        test_data / "esm_datastore/cmip-forcing-qv56.json"
+    )
+    esmds.name = "name"
+    esmds.description = "description"
+    translator = EsgfTranslator(esmds, CORE_COLUMNS)
+    df = translator.translate(groupby)
+    assert len(df) == n_entries
+
+
+@pytest.mark.parametrize(
+    "groupby, n_entries",
+    [
+        (None, 7),
+        (["variable"], 6),
+        (
+            ["frequency"],
+            1,
+        ),
+        (["model"], 6),
+        (["realm"], 1),
+    ],
+)
+def test_esmvaltoolTranslator(test_data, groupby, n_entries):
+    """Test ESMValTool translator"""
+    esmds = intake.open_esm_datastore(test_data / "esm_datastore/access-ct11.json")
+    esmds.name = "name"
+    esmds.description = "description"
+    df = EsmValToolTranslator(esmds, CORE_COLUMNS).translate(groupby)
+    assert len(df) == n_entries
+
+
+def test_translator_failure(test_data):
+    esmds = intake.open_esm_datastore(test_data / "esm_datastore/narclim2-zz63.json")
+    esmds.name = "name"
+    esmds.description = "description"
+    translator = NarclimTranslator(esmds, CORE_COLUMNS)
+
+    default = DefaultTranslator(esmds, CORE_COLUMNS)
+
+    translator.set_dispatch(
+        input_name="dud_name",
+        core_colname="model",
+        func=default._model_translator,
+    )
+
+    with pytest.raises(KeyError) as excinfo:
+        translator.translate()
+
+    assert (
+        "Unable to translate 'model' column with translator 'DefaultTranslator'"
+        in str(excinfo.value)
+    )
+
+    @trace_failure
+    def _(x: int) -> int:
+        return x
+
+    with pytest.raises(TypeError) as excinfo:
+        _(1)
+
+    assert "Decorator can only be applied to translator class methods" in str(
+        excinfo.value
+    )
